requirements.txt
allure-pytest==2.10.0
allure-python-commons==2.10.0
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.1.0
cryptography==36.0.1
execnet==1.9.0
h11==0.12.0
idna==3.3
iniconfig==1.1.1
outcome==1.1.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.9
pytest==7.1.2
pytest-forked==1.4.0
pytest-multithreading-allure==1.0.5
pytest-xdist==2.5.0
python-dotenv==0.20.0
requests==2.28.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.64.0
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
webdriver-manager==3.8.3
wsproto==1.0.0
